New Footer contains StitchTakes player, chapter comments, and Compile button
