+++ src/css/takes.css
@@ -4,7 +4,6 @@
 
 .StickyFooter {
     position: sticky;
-    /*background-color:#B65459;*/
     bottom: 0;
     padding: 0;
